--- src/instrumentation/helper.go
@@ -1,30 +1,0 @@
-/*
-Copyright 2024.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-	http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package instrumentation
-
-import (
-	corev1 "k8s.io/api/core/v1"
-)
-
-// Calculate if we already inject InitContainers.
-func isInitContainerMissing(pod corev1.Pod) bool {
-	for _, initContainer := range pod.Spec.InitContainers {
-		if initContainer.Name == initContainerName {
-			return false
-		}
-	}
-	return true
-}
